IRIS-4246 | refactor utility methods
