+++ src/test/java/com/wikia/webdriver/common/core/helpers/ForgottenPasswordUserPool.java
@@ -1,0 +1,19 @@
+package com.wikia.webdriver.common.core.helpers;
+
+
+import com.wikia.webdriver.common.core.configuration.Configuration;
+import com.wikia.webdriver.common.properties.Credentials;
+
+public class ForgottenPasswordUserPool {
+
+  private static final Credentials CREDENTIALS = Configuration.getCredentials();
+
+  public static UserWithEmail user1 = new UserWithEmail(User.FORGOTTEN_PASSWORD,
+    CREDENTIALS.forgottenPasswordEmail1Address,
+    CREDENTIALS.forgottenPasswordEmail1Password);
+
+  public static UserWithEmail user2 = new UserWithEmail(User.FORGOTTEN_PASSWORD_SPACES,
+    CREDENTIALS.forgottenPasswordEmail2Address,
+    CREDENTIALS.forgottenPasswordEmail2Password);
+
+}
