+++ src/test/java/com/wikia/webdriver/common/core/helpers/UserWithEmail.java
@@ -1,0 +1,21 @@
+package com.wikia.webdriver.common.core.helpers;
+
+import lombok.AllArgsConstructor;
+import lombok.Value;
+
+@AllArgsConstructor
+@Value
+public class UserWithEmail {
+
+  private User user;
+  private String email;
+  private String emailPassword;
+
+  public String getUsername() {
+    return user.getUserName();
+  }
+
+  public String getPassword() {
+    return user.getPassword();
+  }
+}
